Spec 0013 — Move Base/Telemetry* → Telemetry/
TelemetryService and TelemetryInsights move to a new Telemetry/
namespace, with TelemetryInsights renamed to Insights (the surrounding
namespace already supplies the Telemetry context). The
TelemetryServiceInterface and ProvidesTelemetryInsights contracts land
at Contracts/ alongside the other interfaces.
