+++ packages/core/src/Contracts/TelemetryService.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Lunar\Core\Base;
+namespace Lunar\Core\Contracts;
 
-interface TelemetryServiceInterface
+interface TelemetryService
 {
     public function optOut(): void;
