+++ packages/core/src/Facades/Telemetry.php
@@ -3,8 +3,7 @@
 namespace Lunar\Core\Facades;
 
 use Illuminate\Support\Facades\Facade;
-use Lunar\Core\Base\TelemetryService;
-use Lunar\Core\Base\TelemetryServiceInterface;
+use Lunar\Core\Contracts\TelemetryService;
 
 /**
  * @method static void optOut()
@@ -13,12 +12,12 @@ use Lunar\Core\Base\TelemetryServiceInterface;
  * @method static bool shouldRun()
  * @method static void run()
  *
- * @see TelemetryService
+ * @see \Lunar\Core\Telemetry\TelemetryService
  */
 class Telemetry extends Facade
 {
     public static function getFacadeAccessor(): string
     {
-        return TelemetryServiceInterface::class;
+        return TelemetryService::class;
     }
 }
