+++ packages/upgrade/src/Rector/LunarSetList.php
@@ -85,7 +85,7 @@ final class LunarSetList
         'Lunar\\Base\\PaymentManagerInterface' => 'Lunar\\Core\\Base\\PaymentManagerInterface',
         'Lunar\\Base\\PaymentTypeInterface' => 'Lunar\\Core\\Base\\PaymentTypeInterface',
         'Lunar\\Base\\PricingManagerInterface' => 'Lunar\\Core\\Base\\PricingManagerInterface',
-        'Lunar\\Base\\ProvidesTelemetryInsights' => 'Lunar\\Core\\Base\\ProvidesTelemetryInsights',
+        'Lunar\\Base\\ProvidesTelemetryInsights' => 'Lunar\\Core\\Contracts\\ProvidesTelemetryInsights',
         'Lunar\\Base\\Purchasable' => 'Lunar\\Core\\Base\\Purchasable',
         'Lunar\\Base\\ShippingManifest' => 'Lunar\\Core\\Manifests\\ShippingManifest',
         'Lunar\\Base\\ShippingManifestInterface' => 'Lunar\\Core\\Base\\ShippingManifestInterface',
@@ -95,9 +95,9 @@ final class LunarSetList
         'Lunar\\Base\\StorefrontSessionInterface' => 'Lunar\\Core\\Base\\StorefrontSessionInterface',
         'Lunar\\Base\\TaxDriver' => 'Lunar\\Core\\Base\\TaxDriver',
         'Lunar\\Base\\TaxManagerInterface' => 'Lunar\\Core\\Base\\TaxManagerInterface',
-        'Lunar\\Base\\TelemetryInsights' => 'Lunar\\Core\\Base\\TelemetryInsights',
-        'Lunar\\Base\\TelemetryService' => 'Lunar\\Core\\Base\\TelemetryService',
-        'Lunar\\Base\\TelemetryServiceInterface' => 'Lunar\\Core\\Base\\TelemetryServiceInterface',
+        'Lunar\\Base\\TelemetryInsights' => 'Lunar\\Core\\Telemetry\\Insights',
+        'Lunar\\Base\\TelemetryService' => 'Lunar\\Core\\Telemetry\\TelemetryService',
+        'Lunar\\Base\\TelemetryServiceInterface' => 'Lunar\\Core\\Contracts\\TelemetryService',
         'Lunar\\Base\\Traits\\CachesProperties' => 'Lunar\\Core\\Base\\Traits\\CachesProperties',
         'Lunar\\Base\\Traits\\CanScheduleAvailability' => 'Lunar\\Core\\Base\\Traits\\CanScheduleAvailability',
         'Lunar\\Base\\Traits\\HasAttributes' => 'Lunar\\Core\\Base\\Traits\\HasAttributes',
